+++ CHANGELOG.md
@@ -2,14 +2,15 @@
 
 ### Features
 
+-   Added defined in links for classes, enums, #180.
 -   Added support for `*.ghe.com` and `*.github.us` GitHub enterprise domains for source links, #2001.
 -   Expose `Converter.parseRawComment` for plugins to parse additional markdown files, #2004.
--   Added defined in links for classes, enums #180.
 
 ### Bug Fixes
 
--   Fixed missing `sources` property on signature reflections #1996.
+-   TypeDoc will no longer emit a warning for `{@link}` containing a URL, #1980.
 -   `excludeNotDocumented` will no longer remove functions/methods/accessors which are documented, #1994.
+-   Fixed missing `sources` property on signature reflections #1996.
 
 ### Thanks!