+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+## v0.23.8 (2022-07-17)
+
 ### Features
 
 -   Added defined in links for classes, enums, #180.